+++ .config/.cprc.json
@@ -1,3 +1,4 @@
 {
-  "version": "5.25.8"
+  "version": "7.9.0",
+  "features": {}
 }
