+++ package.json
@@ -8,11 +8,11 @@
     "e2e": "playwright test",
     "e2e:debug": "npx playwright test --ui",
     "generate-release-notes": "PREV_TAG=$(git tag | tail -n 1) && gh api --method POST /repos/grafana/iot-sitewise-datasource/releases/generate-notes -f tag_name=v${npm_package_version} -f target_commitish=main -f previous_tag_name=${PREV_TAG} | jq -r .body",
-    "lint": "eslint --cache .",
+    "lint": "eslint --cache --ignore-path ./.gitignore --ext .js,.jsx,.ts,.tsx .",
     "lint:fix": "yarn run lint --fix && prettier --write --list-different .",
     "server": "docker compose up --build",
     "server:dev": "DEVELOPMENT=true yarn server",
-    "sign": "npx --yes @grafana/sign-plugin@latest",
+    "sign": "sign-plugin",
     "spellcheck": "cspell -c cspell.config.json \"**/*.{ts,tsx,js,go,md,mdx,yml,yaml,json,scss,css}\"",
     "test": "jest --watch --onlyChanged",
     "test:ci": "jest --passWithNoTests --maxWorkers 4",
@@ -25,11 +25,12 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@emotion/css": "11.13.5",
-    "@grafana/data": "^12.1.0",
+    "@grafana/data": "^12.2.0",
+    "@grafana/i18n": "^12.2.0",
     "@grafana/plugin-ui": "^0.13.0",
-    "@grafana/runtime": "^12.1.0",
-    "@grafana/schema": "^12.1.0",
-    "@grafana/ui": "^12.1.0",
+    "@grafana/runtime": "^12.2.0",
+    "@grafana/schema": "^12.2.0",
+    "@grafana/ui": "^12.2.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-router-dom": "^6.22.0",
@@ -43,6 +44,7 @@
     "@grafana/aws-sdk": "0.10.2",
     "@grafana/eslint-config": "^9.0.0",
     "@grafana/plugin-e2e": "3.4.13",
+    "@grafana/sign-plugin": "^3.2.2",
     "@grafana/tsconfig": "^2.0.1",
     "@playwright/test": "1.58.2",
     "@stylistic/eslint-plugin-ts": "^4.2.0",
@@ -56,6 +58,8 @@
     "@types/jest": "^30.0.0",
     "@types/lodash": "^4.17.21",
     "@types/node": "24.10.8",
+    "@types/react": "^18.3.0",
+    "@types/react-dom": "^18.3.0",
     "@types/semver": "^7.7.1",
     "@typescript-eslint/eslint-plugin": "^8.56.1",
     "@typescript-eslint/parser": "^8.56.1",
@@ -65,7 +69,6 @@
     "dotenv": "^17.3.1",
     "eslint": "^10.0.0",
     "eslint-config-prettier": "^10.1.5",
-    "eslint-plugin-deprecation": "^3.0.0",
     "eslint-plugin-jsdoc": "^62.7.1",
     "eslint-plugin-prettier": "^5.5.4",
     "eslint-plugin-react": "^7.37.5",
